gcc 6.1 fixes

+++ src/align/align_sequence_anchored.cpp
@@ -112,7 +112,7 @@ void rank_subjects(vector<Subject_seq> &subjects, vector<local_trace_point> &tp)
 		score = (double)subjects[0].filter_score * (1.0 - config.toppercent / 100.0) * config.rank_ratio;
 	}
 	else {
-		size_t min_idx = std::min(subjects.size(), config.max_alignments);
+		size_t min_idx = std::min(subjects.size(), (size_t)config.max_alignments);
 		score = (double)subjects[min_idx - 1].filter_score * config.rank_ratio;
 	}
